refactor SendContainer to class
